# Performance optimizations: Regex caching and modern JS features